requirements.txt
Flask
Flask-SQLAlchemy
Flask-WTF
Flask-Migrate
SQLAlchemy
pytest
pytest-flask
requests
mysql-connector-python
flask-marshmallow 
marshmallow-sqlalchemy